chore: rename files and folders
