+++ src/components/UserProfile/SideWidget/__test__/sideWidget.spec.js
