+++ src/components/UserProfile/SideWidget/side-widget.scss
