+++ src/reducers/index.js
@@ -11,13 +11,13 @@ import articlesResponse from './homeReducers/home';
 import articleReducers from './articleReducers/articleReducers';
 import categoryReducer from './articleReducers/categoryReducer';
 import imageUploadReducers from './articleReducers/imageUploadReducers';
-import profile from './userProfileReducers/Profile';
+import profile from './userProfileReducers/profile';
 import follow from './reactionReducers/follow';
 import commentReducers from './commentReducers/comment';
 import fetchAverageRating from './ratingsReducer/fetchAveargeRatings';
 import postRating from './ratingsReducer/postRatings';
 import fetchUserRatings from './ratingsReducer/fetchUserRatings';
-import statistics from './userProfileReducers/Statistics';
+import statistics from './userProfileReducers/statistics';
 
 const rootReducer = combineReducers({
   signup,
