+++ src/reducers/userProfileReducers/__test__/getArticle.spec.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-len */
 import * as type from '../../../actions/userAction/actionTypes';
-import getArticleReducer, { initialState as defaultState } from '../Profile';
+import getArticleReducer, { initialState as defaultState } from '../profile';
 import updateObject from '../../../helpers/store/utility';
 
 describe('Get Article Reducer', () => {
